backend/about/serializers.py
# serializers.py
from rest_framework import serializers
from .models import AboutMe, ContactInquiry

class AboutMeSerializer(serializers.ModelSerializer):
    class Meta:
        model = AboutMe
        fields = '__all__'

class ContactInquirySerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactInquiry
        fields = '__all__'